refactor: Expose log in pkg
