+++ cmd/commands/check.go
@@ -2,7 +2,7 @@ package commands
 
 import (
 	"github.com/caffeine-addictt/waku/internal/errors"
-	"github.com/caffeine-addictt/waku/internal/log"
+	"github.com/caffeine-addictt/waku/pkg/log"
 	"github.com/caffeine-addictt/waku/internal/template"
 	"github.com/spf13/cobra"
 )
