+++ cmd/commands/healthcheck.go
@@ -3,7 +3,7 @@ package commands
 import (
 	"github.com/caffeine-addictt/waku/internal/errors"
 	"github.com/caffeine-addictt/waku/internal/git"
-	"github.com/caffeine-addictt/waku/internal/log"
+	"github.com/caffeine-addictt/waku/pkg/log"
 	"github.com/spf13/cobra"
 )
