+++ internal/git/check.go
@@ -4,7 +4,7 @@ import (
 	"net/url"
 	"os"
 
-	"github.com/caffeine-addictt/waku/internal/log"
+	"github.com/caffeine-addictt/waku/pkg/log"
 )
 
 type UrlType string
