+++ internal/license/store.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 
 	"github.com/caffeine-addictt/waku/cmd/options"
-	"github.com/caffeine-addictt/waku/internal/log"
+	"github.com/caffeine-addictt/waku/pkg/log"
 	"github.com/caffeine-addictt/waku/pkg/version"
 	"github.com/goccy/go-json"
 )
